Take a signed width in g_utf8_fit()
A width computed below zero - a column position resolved against a
screen narrower than the layout assumed - reached the padding function
as an enormous unsigned count and threw std::length_error, which
nothing catches. Only three of its callers guarded against that. The
function takes a signed width now and prints nothing for a negative
one.

+++ goldlib/gall/gutf8.h
@@ -264,7 +264,7 @@ std::string g_utf8_truncate(const std::string& s, size_t maxcols);
 //  bytes - so columns stopped lining up as soon as the text was not
 //  single-byte.
 
-std::string g_utf8_fit(const char* p, size_t cols);
+std::string g_utf8_fit(const char* p, int cols);
 
 
 //  ------------------------------------------------------------------
